% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a.Mi_WHEAT}
\alias{a.Mi_WHEAT}
\title{a.Mi_WHEAT}
\format{
An object of class \code{numeric} of length 1.
}
\source{
Wang, X. and Mauzerall, D.L., 2004. Characterizing distributions of surface ozone and its impact on grain production in China, Japan and South Korea: 1990 and 2020. Atmospheric Environment, 38(26), pp.4383-4402.
\dontrun{
 library(rfasst);
 rfasst::a.Mi_WHEAT
}
}
\usage{
a.Mi_WHEAT
}
\description{
"a" coefficient for Mi-Wheat
}
\keyword{datasets}
