% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adj_rus}
\alias{adj_rus}
\title{adj_rus}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\source{
TM5-FASST
}
\usage{
adj_rus
}
\description{
Shares to distribute emissions of different species between Russia Eastern (RUE) and Western (RUS)
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::adj_rus
}
}
\keyword{datasets}
