% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_years}
\alias{all_years}
\title{all_years}
\format{
An object of class \code{character} of length 11.
}
\usage{
all_years
}
\description{
Years to be analyzed: c('2005','2010','2020','2030','2040','2050','2060','2070','2080','2090','2100')
\dontrun{
 library(rfasst);
 rfasst::all_years
}
}
\keyword{datasets}
