% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b.Mi_RICE}
\alias{b.Mi_RICE}
\title{b.Mi_RICE}
\format{
An object of class \code{numeric} of length 1.
}
\source{
Wang, X. and Mauzerall, D.L., 2004. Characterizing distributions of surface ozone and its impact on grain production in China, Japan and South Korea: 1990 and 2020. Atmospheric Environment, 38(26), pp.4383-4402.
\dontrun{
 library(rfasst);
 rfasst::b.Mi_RICE
}
}
\usage{
b.Mi_RICE
}
\description{
"b" coefficient for Mi-Rice
}
\keyword{datasets}
