% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_dongetal2021}
\alias{beta_dongetal2021}
\title{beta_dongetal2021}
\format{
An object of class \code{numeric} of length 1.
}
\source{
OECD
\dontrun{
 library(rfasst);
 rfasst::beta_dongetal2021
}
}
\usage{
beta_dongetal2021
}
\description{
Beta for damages in GDP growth based on Dong et al 2021
}
\keyword{datasets}
