% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_health.R
\name{calc_daly_pm25}
\alias{calc_daly_pm25}
\title{calc_daly_pm25}
\source{
Institute for Health Metrics and Evaluation (http://www.healthdata.org/)
}
\usage{
calc_daly_pm25()
}
\value{
DALY-to-Mortality ratios for TM5-FASST regions for all years and PM2.5-related causes (ALRI, COPD, LC, IHD, STROKE).The list of countries that form each region and the full name of the region can be found in Table S2.2 in the TM5-FASST documentation paper: Van Dingenen, R., Dentener, F., Crippa, M., Leitao, J., Marmer, E., Rao, S., Solazzo, E. and Valentini, L., 2018. TM5-FASST: a global atmospheric source-receptor model for rapid impact analysis of emission changes on air quality and short-lived climate pollutants. Atmospheric Chemistry and Physics, 18(21), pp.16173-16211.
}
\description{
Get the DALY-to-Mortality ratios used to estimate the Disability Adjusted Life Years attributable to fine particulate matter (PM2.5) exposure
}
\keyword{DALYs,}
\keyword{PM2.5}
