% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_health.R
\name{calc_daly_tot}
\alias{calc_daly_tot}
\title{calc_daly_tot}
\source{
Institute for Health Metrics and Evaluation (http://www.healthdata.org/)
}
\usage{
calc_daly_tot()
}
\value{
DALYs for TM5-FASST regions for all years and causes (ALRI, COPD, LC, IHD, STROKE)
}
\description{
Get the DALY-to-Mortality ratios used to estimate the Disability Adjusted Life Years.
}
\keyword{DALYs}
