% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{calc_gdp_pc_reg}
\alias{calc_gdp_pc_reg}
\title{calc_gdp_pc_reg}
\source{
https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome
}
\usage{
calc_gdp_pc_reg(ssp = "SSP2")
}
\arguments{
\item{ssp}{Set the ssp narrative associated to the GCAM scenario. c("SSP1","SSP2","SSP3","SSP4","SSP5"). By default is SSP2}
}
\value{
GDP_pc for TM5-FASST regions for all years
}
\description{
Get GDP_pc from the SSP database (SSP_database_v9) for the economic assessment of the health impacts.
To be consistent we make use of the IIASA Model/scenarios. rfasst regions considered.
}
\keyword{GDP}
\keyword{socioeconomics,}
