% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_health.R
\name{calc_mort_rates}
\alias{calc_mort_rates}
\title{calc_mort_rates}
\source{
https://www.who.int/healthinfo/global_burden_disease/cod_2008_sources_methods.pdf
}
\usage{
calc_mort_rates()
}
\value{
Baseline mortality rates for TM5-FASST regions for all years and causes (ALRI, COPD, LC, IHD, STROKE)
}
\description{
Get cause-specific baseline mortalities from stroke, ischemic heart disease (IHD), chronic obstructive pulmonary disease (COPD), acute lower respiratory illness diseases (ALRI) and lung cancer (LC).
}
\keyword{Baseline}
\keyword{mortality}
\keyword{rates}
