% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{calc_pop}
\alias{calc_pop}
\title{calc_pop}
\source{
https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome
}
\usage{
calc_pop(ssp = "SSP2")
}
\arguments{
\item{ssp}{Set the ssp narrative associated to the GCAM scenario. c("SSP1","SSP2","SSP3","SSP4","SSP5"). By default is SSP2}
}
\value{
Population and population shares (<5Y; >30Y) for TM5-FASST regions for all years
}
\description{
Get population data and shares of population under 5 Years and above 30 Years from the SSP database (SSP_database_v9).To be consistent we make use of the IIASA-WIC Model/scenarios. Given that IIASA-WIC does not report data for Taiwan, we use data from "OECD_Env-Growth" for this region.
}
\keyword{population}
\keyword{socioeconomics,}
