% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_agriculture.R
\name{calc_prod_gcam}
\alias{calc_prod_gcam}
\title{calc_prod_gcam}
\source{
GCAM
}
\usage{
calc_prod_gcam(
  db_path,
  query_path,
  db_name,
  prj_name,
  scen_name,
  queries,
  saveOutput = T
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{saveOutput}{Writes the emission files.By default=T}
}
\value{
Future agricultural production levels from GCAM for each GCAM region for all years
}
\description{
Extract future agricultural production levels from GCAM
}
\keyword{agriculture,}
\keyword{module_4,}
\keyword{production}
