% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{calc_rr_updt}
\alias{calc_rr_updt}
\title{calc_rr_updt}
\source{
"Assessment of air pollution health co-benefits of Net-zero climate policies" Masther's thesis by CRB
}
\usage{
calc_rr_updt()
}
\value{
RR for a series of PM2.5 levels
}
\description{
Function to calculate RR tables using updated epidemiological functions
}
