% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_o3}
\alias{cf_o3}
\title{cf_o3}
\format{
An object of class \code{numeric} of length 1.
}
\source{
Jerrett, M., Burnett, R.T., Pope III, C.A., Ito, K., Thurston, G., Krewski, D., Shi, Y., Calle, E. and Thun, M., 2009. Long-term ozone exposure and mortality. New England Journal of Medicine, 360(11), pp.1085-1095.
\dontrun{
 library(rfasst);
 rfasst::cf_o3
}
}
\usage{
cf_o3
}
\description{
Counterfactual threshold for ozone (Jerret et al 2009)
}
\keyword{datasets}
