% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch4_htap_pert}
\alias{ch4_htap_pert}
\title{ch4_htap_pert}
\format{
An object of class \code{numeric} of length 1.
}
\source{
Fiore, A.M., West, J.J., Horowitz, L.W., Naik, V. and Schwarzkopf, M.D., 2008. Characterizing the tropospheric ozone response to methane emission controls and the benefits to climate and air quality. Journal of Geophysical Research: Atmospheres, 113(D8).
\dontrun{
 library(rfasst);
 rfasst::ch4_htap_pert
}
}
\usage{
ch4_htap_pert
}
\description{
Normalized CH4-O3 relation from Fiore et al (2008)
}
\keyword{datasets}
