% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{check_byu}
\alias{check_byu}
\title{check_byu}
\usage{
check_byu(data)
}
\arguments{
\item{data:}{dataset}
}
\value{
dataset with the year 2020 (corresponding to the estimated 2021 if present).
}
\description{
Check if the data contains GCAM results after the BYU, i.e., with the year 2021
instead of the year 2020, and change the year to match the socioeconomic preloaded data
}
