% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{check_gcamversion}
\alias{check_gcamversion}
\title{check_gcamversion}
\usage{
check_gcamversion(tmp, gcam_eur)
}
\arguments{
\item{tmp}{Dummy dataset (any query from the prj file) to check if the BYU (and regions' update) is implemented in the GCAM considered output.}

\item{gcam_eur}{If set to T, considers the GCAM-Europe regions.}
}
\value{
list of GCAM version region-dependent constants
}
\description{
Check if the data contains GCAM results from GCAM8.2 or newer versions to arrange
the regions (e.g., Ukraine)
}
