% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{clean_pkg_outputs}
\alias{clean_pkg_outputs}
\title{clean_pkg_outputs}
\usage{
clean_pkg_outputs()
}
\value{
List with the variables to be erased. Type "rm(list = clean_pkg_outputs())" to remove them
}
\description{
List of global variables generated by the pkg to be erased
}
