% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coef.AOT_MAIZE}
\alias{coef.AOT_MAIZE}
\title{coef.AOT_MAIZE}
\format{
An object of class \code{numeric} of length 1.
}
\source{
Mills, G., Buse, A., Gimeno, B., Bermejo, V., Holland, M., Emberson, L. and Pleijel, H., 2007. A synthesis of AOT40-based response functions and critical levels of ozone for agricultural and horticultural crops. Atmospheric Environment, 41(12), pp.2630-2643.
\dontrun{
 library(rfasst);
 rfasst::coef.AOT_MAIZE
}
}
\usage{
coef.AOT_MAIZE
}
\description{
Coefficient for AOT40-Maize
}
\keyword{datasets}
