% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coef.Mi_RICE}
\alias{coef.Mi_RICE}
\title{coef.Mi_RICE}
\format{
An object of class \code{numeric} of length 1.
}
\source{
Wang, X. and Mauzerall, D.L., 2004. Characterizing distributions of surface ozone and its impact on grain production in China, Japan and South Korea: 1990 and 2020. Atmospheric Environment, 38(26), pp.4383-4402.
\dontrun{
 library(rfasst);
 rfasst::coef.Mi_RICE
}
}
\usage{
coef.Mi_RICE
}
\description{
Coefficient for Mi-Rice
}
\keyword{datasets}
