% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONV_1975_2010}
\alias{CONV_1975_2010}
\title{CONV_1975_2010}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CONV_1975_2010
}
\description{
1975-2010 deflator
\dontrun{
 library(rfasst);
 rfasst::CONV_1975_2010
}
}
\keyword{datasets}
