% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONV_BIL}
\alias{CONV_BIL}
\title{CONV_BIL}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CONV_BIL
}
\description{
Transform $Billion to $
\dontrun{
 library(rfasst);
 rfasst::CONV_BIL
}
}
\keyword{datasets}
