% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONV_MIL}
\alias{CONV_MIL}
\title{CONV_MIL}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CONV_MIL
}
\description{
Transform $Million to $
\dontrun{
 library(rfasst);
 rfasst::CONV_MIL
}
}
\keyword{datasets}
