% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONV_OC_POM}
\alias{CONV_OC_POM}
\title{CONV_OC_POM}
\format{
An object of class \code{numeric} of length 1.
}
\source{
Kanakidou, M., Seinfeld, J.H., Pandis, S.N., Barnes, I., Dentener, F.J., Facchini, M.C., Dingenen, R.V., Ervens, B., Nenes, A.N.C.J.S.E., Nielsen, C.J. and Swietlicki, E., 2005. Organic aerosol and global climate modelling: a review. Atmospheric Chemistry and Physics, 5(4), pp.1053-1123.
\dontrun{
 library(rfasst);
 rfasst::CONV_OC_POM
}
}
\usage{
CONV_OC_POM
}
\description{
Transform OC to POM
}
\keyword{datasets}
