% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONV_TG_T}
\alias{CONV_TG_T}
\title{CONV_TG_T}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CONV_TG_T
}
\description{
Unit converter: Teragram to tonne
\dontrun{
 library(rfasst);
 rfasst::CONV_TG_T
}
}
\keyword{datasets}
