% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_iso}
\alias{country_iso}
\title{country_iso}
\format{
An object of class \code{data.frame} with 249 rows and 2 columns.
}
\usage{
country_iso
}
\description{
Mapping of countries to iso3 codes
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::country_iso
}
}
\keyword{datasets}
