% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CROP_ANALYSIS}
\alias{CROP_ANALYSIS}
\title{CROP_ANALYSIS}
\format{
An object of class \code{character} of length 12.
}
\usage{
CROP_ANALYSIS
}
\description{
Crops that are included in the analysis
\dontrun{
 library(rfasst);
 rfasst::CROP_ANALYSIS
}
}
\keyword{datasets}
