% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctry_nuts3_codes}
\alias{ctry_nuts3_codes}
\title{ctry_nuts3_codes}
\format{
.csv
}
\usage{
ctry_nuts3_codes
}
\description{
A mapping table that links ISO2, ISO3, and NUTS3 codes. For regions outside Europe, the NUTS3 code corresponds to the ISO3 code.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::ctry_nuts3_codes
}
}
\keyword{datasets}
