% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctry_nuts_sf}
\alias{ctry_nuts_sf}
\title{ctry_nuts_sf}
\format{
sf
}
\source{
Processing of territorial-units-statistics. Details in saveDataFiles.R
}
\usage{
ctry_nuts_sf
}
\description{
sf object for the entire World containing the NUTS3 region if avilable (Europe) or the country area otherwise
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::ctry_nuts_sf
}
}
\keyword{datasets}
