% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.gcam.commod.o3}
\alias{d.gcam.commod.o3}
\title{d.gcam.commod.o3}
\format{
An object of class \code{data.frame} with 39 rows and 2 columns.
}
\source{
Own assumptions
\dontrun{
 library(rfasst);
 rfasst::d.gcam.commod.o3
}
}
\usage{
d.gcam.commod.o3
}
\description{
O3 to GCAM commodities (based on their carbon fixation pathways; C3 and C4 categories)
}
\keyword{datasets}
