% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.ha}
\alias{d.ha}
\title{d.ha}
\format{
An object of class \code{data.frame} with 1109 rows and 4 columns.
}
\source{
GFDL-NOAA
\dontrun{
 library(rfasst);
 rfasst::d.ha
}
}
\usage{
d.ha
}
\description{
Harvested area by crop for weights to map O3 crops to GCAM commodities
}
\keyword{datasets}
