% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.iso}
\alias{d.iso}
\title{d.iso}
\format{
An object of class \code{data.frame} with 238 rows and 4 columns.
}
\usage{
d.iso
}
\description{
Countries to GCAM regions
\dontrun{
 library(rfasst);
 rfasst::d.iso
}
}
\keyword{datasets}
