% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.iso_EUR}
\alias{d.iso_EUR}
\title{d.iso_EUR}
\format{
An object of class \code{data.frame} with 242 rows and 4 columns.
}
\usage{
d.iso_EUR
}
\description{
Countries to GCAM-Europe regions
\dontrun{
 library(rfasst);
 rfasst::d.iso_EUR
}
}
\keyword{datasets}
