% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.iso_vlt8.2}
\alias{d.iso_vlt8.2}
\title{d.iso_vlt8.2}
\format{
An object of class \code{data.frame} with 242 rows and 4 columns.
}
\usage{
d.iso_vlt8.2
}
\description{
Countries to GCAM (version < 8.2) regions
\dontrun{
 library(rfasst);
 rfasst::d.iso_vlt8.2
}
}
\keyword{datasets}
