% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.weight.gcam_vgt8.2}
\alias{d.weight.gcam_vgt8.2}
\title{d.weight.gcam_vgt8.2}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1215 rows and 4 columns.
}
\source{
Own assumptions
\dontrun{
 library(rfasst);
 rfasst::d.weight.gcam_vgt8.2
}
}
\usage{
d.weight.gcam_vgt8.2
}
\description{
O3 to GCAM (version >= 8.2) commodities (based on their carbon fixation pathways; C3 and C4 categories)
}
\keyword{datasets}
