% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{data_query}
\alias{data_query}
\title{data_query}
\usage{
data_query(
  type,
  db_path,
  db_name,
  prj_name,
  scenarios,
  query_path = "inst/extdata",
  queries = "queries_rfasst_nonCO2.xml"
)
}
\arguments{
\item{query_path}{Path to the query file}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{db_path:}{path of the database}

\item{db_name:}{name of the database}

\item{prj_name:}{name of the project}

\item{scenarios:}{name of the scenarios to be considered}

\item{type:}{either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'}
}
\value{
dataframe with the data from the query
}
\description{
Add nonCO2 large queries
}
