% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dis}
\alias{dis}
\title{dis}
\format{
An object of class \code{character} of length 5.
}
\source{
TM5-FASST
\dontrun{
 library(rfasst);
 rfasst::dis
}
}
\usage{
dis
}
\description{
List of diseases for reading relative risk
}
\keyword{datasets}
