% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{exampleFunction}
\alias{exampleFunction}
\title{Details: Add any details or general description about this function here.
Details: Add any details or general description about this function here.}
\usage{
exampleFunction(param1 = 1, param2 = 2)
}
\arguments{
\item{param1}{param1 Description}

\item{param2}{param2 Description}
}
\description{
Details: Add any details or general description about this function here.
Details: Add any details or general description about this function here.
}
\examples{
\donttest{
    library(exampleRepo)
    exampleFunction(3,4)
    }
}
