% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fasst_reg}
\alias{fasst_reg}
\title{fasst_reg}
\format{
An object of class \code{data.frame} with 249 rows and 2 columns.
}
\source{
TM5-FASST
}
\usage{
fasst_reg
}
\description{
Mapping of countries to TM5-FASST regions
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::fasst_reg
}
}
\keyword{datasets}
