% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GCAM_reg}
\alias{GCAM_reg}
\title{GCAM_reg}
\format{
An object of class \code{data.frame} with 234 rows and 3 columns.
}
\source{
GCAM
}
\usage{
GCAM_reg
}
\description{
Mapping of countries to GCAM regions
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::GCAM_reg
}
}
\keyword{datasets}
