% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GCAM_reg_vgt8.2}
\alias{GCAM_reg_vgt8.2}
\title{GCAM_reg_vgt8.2}
\format{
An object of class \code{data.frame} with 235 rows and 3 columns.
}
\source{
GCAM
}
\usage{
GCAM_reg_vgt8.2
}
\description{
Mapping of countries to GCAM (version >= 8.2) regions
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::GCAM_reg_vgt8.2
}
}
\keyword{datasets}
