% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_eu_2005}
\alias{gdp_eu_2005}
\title{gdp_eu_2005}
\format{
An object of class \code{numeric} of length 1.
}
\source{
OECD
\dontrun{
 library(rfasst);
 rfasst::gdp_eu_2005
}
}
\usage{
gdp_eu_2005
}
\description{
Base GDP for EU in 2005
}
\keyword{datasets}
