% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_growth.SSP3}
\alias{gdp_growth.SSP3}
\title{GDP-SSP database}
\format{
.csv
}
\source{
https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome
}
\usage{
gdp_growth.SSP3
}
\description{
GDP growth rates. To be consistent we make use of the IIASA Model/scenarios
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::gdp_growth.SSP3
}
}
\keyword{datasets}
