% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_pc.ctry_nuts3.SSP5}
\alias{gdp_pc.ctry_nuts3.SSP5}
\title{GDP-SSP database}
\format{
.csv
}
\source{
https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome
}
\usage{
gdp_pc.ctry_nuts3.SSP5
}
\description{
Filtered GDP data per SSP5 (SSP_database_v9.csv).
To be consistent we make use of the IIASA Model/scenarios. Data aggregated at NUTS3 level when possible. Otherwise at ISO3 level.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::gdp_pc.ctry_nuts3.SSP5
}
}
\keyword{datasets}
