% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_pc.SSP4}
\alias{gdp_pc.SSP4}
\title{GDP-SSP database}
\format{
.csv
}
\source{
https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome
}
\usage{
gdp_pc.SSP4
}
\description{
Filtered GDP data per SSP4 (SSP_database_v9.csv). To be consistent we make use of the IIASA Model/scenarios
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::gdp_pc.SSP4
}
}
\keyword{datasets}
