% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{if_complex}
\alias{if_complex}
\title{if_complex}
\usage{
if_complex(condition, data1, data2)
}
\arguments{
\item{condition:}{condifiton to be satisfied to assing data1}

\item{data1:}{dataset1}

\item{data2:}{dataset2}
}
\value{
dataset accroding to the given conditon
}
\description{
Do an if_else to assign data1 or data2 to a dataset
}
