% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inc_elas_vsl}
\alias{inc_elas_vsl}
\title{inc_elas_vsl}
\format{
An object of class \code{numeric} of length 1.
}
\source{
Own assumptions
\dontrun{
 library(rfasst);
 rfasst::inc_elas_vsl
}
}
\usage{
inc_elas_vsl
}
\description{
Income elasticity for the Value of Statistical Life (VSL)
}
\keyword{datasets}
