% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{interpLinear}
\alias{interpLinear}
\title{interpLinear}
\usage{
interpLinear(d, y_start, y_end)
}
\arguments{
\item{d}{Data frame to be interpolated}

\item{y_start}{Starting year (start of the decade)}

\item{y_end}{End year (End of the decade)}
}
\description{
Function to interpolate annual values using decade-averages
}
\keyword{interpolate}
