% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{load_prj}
\alias{load_prj}
\title{load_prj}
\usage{
load_prj(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name,
  prj = NULL,
  scen_name,
  queries = "queries_rfasst.xml"
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{prj}{rgcam loaded project}

\item{scen_name}{Vector names of the GCAM scenarios to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}
}
\value{
loaded project
}
\description{
Load/Create prj file
}
