% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_conc.R
\name{m2_get_conc_mi}
\alias{m2_get_conc_mi}
\title{m2_get_conc_mi}
\usage{
m2_get_conc_mi(
  db_path,
  query_path,
  db_name,
  prj_name,
  scen_name,
  queries,
  saveOutput = T,
  map = F
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{saveOutput}{Writes the emission files.By default=T}

\item{map}{Produce the maps. By default=F}
}
\value{
Produce Mi levels for each TM5-FASST regions for all years
}
\description{
Produce Mi concentration levels based on re-scaled emission pathways from module 1. Mi is the the seasonal mean daytime O3 concentration (M7 for the 7-hour mean and M12 for the 12-hour mean)
}
\keyword{(M7}
\keyword{M12)}
\keyword{Mi}
\keyword{and}
\keyword{concentration,}
\keyword{module_2,}
