% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_conc.R
\name{m2_get_conc_o3}
\alias{m2_get_conc_o3}
\title{m2_get_conc_o3}
\usage{
m2_get_conc_o3(
  db_path,
  query_path,
  db_name,
  prj_name,
  scen_name,
  queries,
  saveOutput = T,
  ch4_o3 = T
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{saveOutput}{Writes the emission files.By default=T}

\item{ch4_o3}{Includes the CH4 effect on O3 based on Fiore et al (2008).By default=T}
}
\value{
Produce ozone (O3) levels for each TM5-FASST regions for all years (ppb)
}
\description{
Produce ozone (O3) concentration levels based on re-scaled emission pathways
}
\keyword{O3}
\keyword{concentration,}
\keyword{module_2,}
