% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_health.R
\name{m3_get_mort_pm25}
\alias{m3_get_mort_pm25}
\title{m3_get_mort_pm25}
\source{
Burnett, R.T., Pope III, C.A., Ezzati, M., Olives, C., Lim, S.S., Mehta, S., Shin, H.H., Singh, G., Hubbell, B., Brauer, M. and Anderson, H.R., 2014. An integrated risk function for estimating the global burden of disease attributable to ambient fine particulate matter exposure. Environmental health perspectives, 122(4), pp.397-403.
}
\usage{
m3_get_mort_pm25(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name,
  prj = NULL,
  scen_name,
  queries = "queries_rfasst.xml",
  final_db_year = 2100,
  mort_param = "GBD",
  ssp = "SSP2",
  saveOutput = T,
  map = F,
  anim = T,
  by_sex = F,
  recompute = F,
  gcam_eur = F,
  normalize = F,
  downscale = F,
  saveRaster_grid = F,
  agg_grid = F,
  save_AggGrid = F
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{prj}{rgcam loaded project}

\item{scen_name}{Vector names of the GCAM scenarios to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{final_db_year}{Final year in the GCAM database (this allows to process databases with user-defined "stop periods")}

\item{mort_param}{Method to compute the premature mortality estimates used in the plotting: GBD, GEMM, or FUSION}

\item{ssp}{Set the ssp narrative associated to the GCAM scenario. c("SSP1","SSP2","SSP3","SSP4","SSP5"). By default is SSP2}

\item{saveOutput}{Writes the emission files.By default=T}

\item{map}{Produce the maps. By default=F}

\item{anim}{If set to T, produces multi-year animations. By default=T}

\item{by_sex}{If set to T, the output is stored separately by sex ("Female", "Male", and "Both"). If set to F, (default), only the "Both" category is saved.}

\item{recompute}{If set to T, recomputes the function output. Otherwise, if the output was already computed once, it uses that value and avoids repeating computations. By default=F}

\item{gcam_eur}{If set to T, considers the GCAM-Europe regions. By default=F}

\item{normalize}{Adjust the output to represent the number of deaths per 100K people. By default = F}

\item{downscale}{If set to T, produces gridded PM2.5 outputs and plots By default=F}

\item{saveRaster_grid}{If set to T, writes the raster file with weighted PM25 By default=F}

\item{agg_grid}{Re-aggregate (downscaled) gridded data to any provided geometries (shape file). For the moment, only "NUTS3" and "CTRY" are available}

\item{save_AggGrid}{If set to T, writes the raster file with the reaggregated PM25 By default=F}
}
\value{
Premature mortality attributable to PM2.5 exposure for each TM5-FASST regions for all years (# mortalities).
}
\description{
Produce premature mortality attributable to PM2.5 exposure based on the integrated exposure-response functions (IER) from Burnett et al (2014), consistent with the GBD 2016 study.
}
\keyword{PM2.5}
\keyword{module_3,}
\keyword{mortality,}
\keyword{premature}
