% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3b_health_ecoloss.R
\name{m3_get_pm25_ecoloss_hcl}
\alias{m3_get_pm25_ecoloss_hcl}
\title{m3_get_pm25_ecoloss_hcl}
\source{
Liu, G., Yang, Z., Chen, B. and Ulgiati, S., 2011. Monitoring trends of urban development and environmental impact of Beijing, 1999–2006. Science of the Total Environment, 409(18), pp.3295-3308.
}
\usage{
m3_get_pm25_ecoloss_hcl(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name = NULL,
  prj = NULL,
  rdata_name = NULL,
  scen_name,
  ssp = "SSP2",
  final_db_year = 2100,
  mort_model = "GBD",
  Damage_vsl_range = "VSL_med",
  inc_elas_vsl = 0.8,
  queries = "queries_rfasst.xml",
  saveOutput = T,
  map = F,
  anim = T,
  recompute = F
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{prj}{rgcam loaded project}

\item{rdata_name}{Name of the RData file. It must contain the queries in a list}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{ssp}{Set the ssp narrative associated to the GCAM scenario. c("SSP1","SSP2","SSP3","SSP4","SSP5"). By default is SSP2}

\item{final_db_year}{Final year in the GCAM database (this allows to process databases with user-defined "stop periods")}

\item{mort_model}{Select the health impact model (GBD, GEMM, or FUSION). By default = GBD}

\item{Damage_vsl_range}{Select the VSL to calculate the damages (VSL_med, VSL_low, or VSL_high). By default = VSL_med}

\item{inc_elas_vsl}{Select the VSL to calculate the damages (VSL_med, VSL_low, or VSL_high). By default = VSL_med}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{saveOutput}{Writes the emission files.By default=T}

\item{map}{Produce the maps. By default=F}

\item{anim}{If set to T, produces multi-year animations. By default=T}

\item{recompute}{If set to T, recomputes the function output. Otherwise, if the output was already computed once, it uses that value and avoids repeating computations. By default=F}
}
\value{
Economic damages associated with mortality attributable to PM2.5 exposure for each TM5-FASST regions for all years (Million$2015). The list of countries that form each region and the full name of the region can be found in Table S2.2 in the TM5-FASST documentation paper: Van Dingenen, R., Dentener, F., Crippa, M., Leitao, J., Marmer, E., Rao, S., Solazzo, E. and Valentini, L., 2018. TM5-FASST: a global atmospheric source-receptor model for rapid impact analysis of emission changes on air quality and short-lived climate pollutants. Atmospheric Chemistry and Physics, 18(21), pp.16173-16211.
}
\description{
Produce economic damages associated with premature mortality attributable to PM2.5. The economic valuation is based on the Human Capital Loss approach
}
\keyword{,premature}
\keyword{HCL}
\keyword{PM2.5}
\keyword{module_3,}
\keyword{mortality,}
