% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_health.R
\name{m3_get_yll_o3}
\alias{m3_get_yll_o3}
\title{m3_get_yll_o3}
\usage{
m3_get_yll_o3(
  db_path,
  query_path,
  db_name,
  prj_name,
  scen_name,
  queries,
  ssp = "SSP2",
  saveOutput = T
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{ssp}{Set the ssp narrative associated to the GCAM scenario. c("SSP1","SSP2","SSP3","SSP4","SSP5"). By default is SSP2}

\item{saveOutput}{Writes the emission files.By default=T}
}
\value{
YLLs attributable to O3 exposure for each TM5-FASST regions for all years
}
\description{
Produce YLLs attributable to O3 (M6M) exposure. YLL-to-Mortalities ratios are based on TM5-FASST calculations. Premature mortalities are based on the IER functions from Jerret et al (2009), consistent with the GBD 2016 study.
}
\keyword{O3}
\keyword{YLL,}
\keyword{module_3,}
