% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_health.R
\name{m3_get_yll_pm25}
\alias{m3_get_yll_pm25}
\title{m3_get_yll_pm25}
\usage{
m3_get_yll_pm25(
  db_path,
  query_path,
  db_name,
  prj_name,
  scen_name,
  queries,
  ssp = "SSP2",
  saveOutput = T,
  map = F
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{ssp}{Set the ssp narrative associated to the GCAM scenario. c("SSP1","SSP2","SSP3","SSP4","SSP5"). By default is SSP2}

\item{saveOutput}{Writes the emission files.By default=T}

\item{map}{Produce the maps. By default=F}
}
\value{
YLLs attributable to PM2.5 exposure for each TM5-FASST regions for all years
}
\description{
Produce YLLs attributable to PM2.5 exposure. YLL-to-Mortalities ratios are based on TM5-FASST calculations. Premature mortalities are  based on the integrated exposure-response functions (IER) from Burnett et al (2014), consistent with the GBD 2016 study.
}
\keyword{PM2.5}
\keyword{YLL,}
\keyword{module_3,}
