% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_agriculture.R
\name{m4_get_ryl_mi}
\alias{m4_get_ryl_mi}
\title{m4_get_ryl_mi}
\usage{
m4_get_ryl_mi(
  db_path,
  query_path,
  db_name,
  prj_name,
  scen_name,
  queries,
  saveOutput = T,
  map = F
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{saveOutput}{Writes the emission files.By default=T}

\item{map}{Produce the maps. By default=F}
}
\value{
RYLs for each TM5-FASST regions for all years
}
\description{
Produce Relative Yield Losses (RYLs) based on the Mi indicator for O3 exposure
}
\keyword{Mi}
\keyword{RYLS,}
\keyword{agriculture,}
\keyword{module_4,}
