% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_pol}
\alias{map_pol}
\title{map_pol}
\format{
An object of class \code{character} of length 6.
}
\usage{
map_pol
}
\description{
Indicate the pollutants whose emissions are mapped (if map=T in m1_emissions_rescale)
\dontrun{
 library(rfasst);
 rfasst::map_pol
}
}
\keyword{datasets}
