% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MTC_MTCO2}
\alias{MTC_MTCO2}
\title{MTC_MTCO2}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MTC_MTCO2
}
\description{
Transform MTC to MTCO2
\dontrun{
 library(rfasst);
 rfasst::MTC_MTCO2
}
}
\keyword{datasets}
