% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_pol}
\alias{my_pol}
\title{#' my_pol}
\format{
An object of class \code{data.frame} with 42 rows and 2 columns.
}
\source{
GCAM
\dontrun{
 library(rfasst);
 rfasst::my_pol
}
}
\usage{
my_pol
}
\description{
Information about GCAM and TM5-FASST regions and pollutants and their equivalences.
}
\keyword{datasets}
