% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts_europe_sf}
\alias{nuts_europe_sf}
\title{nuts_europe_sf}
\format{
sf
}
\source{
Processing of territorial-units-statistics. Details in saveDataFiles.R
}
\usage{
nuts_europe_sf
}
\description{
sf object for Europe containing the NUTS3 region code of European continent regions (no overseas)
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::nuts_europe_sf
}
}
\keyword{datasets}
