% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts_sf}
\alias{nuts_sf}
\title{nuts_sf}
\format{
sf
}
\source{
Processing of territorial-units-statistics. Details in saveDataFiles.R
}
\usage{
nuts_sf
}
\description{
sf object for Europe containing the NUTS3 region code
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::nuts_sf
}
}
\keyword{datasets}
