% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perc_pop_rue}
\alias{perc_pop_rue}
\title{perc_pop_rue}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
perc_pop_rue
}
\description{
Percentages to divide population between Russia and Russia Eastern
\dontrun{
 library(rfasst);
 rfasst::perc_pop_rue
}
}
\keyword{datasets}
