% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop.all.ctry_nuts3.str.SSP2}
\alias{pop.all.ctry_nuts3.str.SSP2}
\title{GDP-SSP database}
\format{
.csv
}
\source{
https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome
}
\usage{
pop.all.ctry_nuts3.str.SSP2
}
\description{
Filtered Population data per SSP5 (SSP_database_v9.csv) by age and sex aggregated at NUTS3 level when possible. Otherwise at ISO3 level.
To be consistent we make use of the IIASA Model/scenarios
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::pop.all.ctry_nuts3.str.SSP2
}
}
\keyword{datasets}
