% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop.all.SSP4}
\alias{pop.all.SSP4}
\title{GDP-SSP database}
\format{
.csv
}
\source{
https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome
}
\usage{
pop.all.SSP4
}
\description{
Filtered Population data per SSP4 (SSP_database_v9.csv). To be consistent we make use of the IIASA Model/scenarios
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::pop.all.SSP4
}
}
\keyword{datasets}
