% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.base_aot}
\alias{raw.base_aot}
\title{Base AOT40 conc}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
raw.base_aot
}
\description{
O3 concentration above a threshold of 40 ppbV (AOT40), in the base year (2000) per TM5-FASST region
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.base_aot
}
}
\keyword{datasets}
