% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.base_conc.nat}
\alias{raw.base_conc.nat}
\title{Base Nat PM2.5 concentration}
\format{
.csv
}
\source{
Results from CAMS
}
\usage{
raw.base_conc.nat
}
\description{
Natural Fine Particulate Matter (PM2.5) concentration levels in the base year (2000) per TM5-FASST region
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.base_conc.nat
}
}
\keyword{datasets}
