% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.base_conc}
\alias{raw.base_conc}
\title{Base concentration}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
raw.base_conc
}
\description{
Fine Particulate Matter (PM2.5), and Ozone (O3 and M6M) concentration levels in the base year (2000) per TM5-FASST region
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.base_conc
}
}
\keyword{datasets}
