% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.base_em}
\alias{raw.base_em}
\title{Base emissions}
\format{
.csv
}
\source{
RCP database: https://tntcat.iiasa.ac.at/RcpDb/dsd?Action=htmlpage&page=welcome
}
\usage{
raw.base_em
}
\description{
Emissions in the base year (2000) per TM5-FASST region of the main precursors for Particulate Matter (PM2.5), and Ozone (O3 and M6M) formation. These include Black Carbon (BC), Carbon Dioxide (CO2), Methane (CH4), Nitrogen Dioxide (N2O), Organic Matter (POM), Nitrogen Oxides (NOx), sulphur dioxide (SO2), ammonia (NH3) and non-methane volatile organic compounds (NMVOC or VOC)
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.base_em
}
}
\keyword{datasets}
