% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.base_mi}
\alias{raw.base_mi}
\title{Base Mi conc}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
raw.base_mi
}
\description{
Seasonal mean daytime O3 concentration (M7 for the 7-hour mean and M12 for the 12-hour mean), in the base year (2000) per TM5-FASST region
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.base_mi
}
}
\keyword{datasets}
