% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.daly}
\alias{raw.daly}
\title{raw.daly}
\format{
An object of class \code{data.frame} with 85680 rows and 18 columns.
}
\source{
Institute for Health Metrics and Evaluation (http://www.healthdata.org/)
}
\usage{
raw.daly
}
\description{
Data on Disability Adjusted Life Years (DALYs). Used the latest available data (2019)
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.daly
}
}
\keyword{datasets}
