% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.gdp}
\alias{raw.gdp}
\title{GDP-SSP database}
\format{
.csv
}
\source{
https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome
}
\usage{
raw.gdp
}
\description{
Filtered GDP data per SSP (SSP_database_v9.csv). To be consistent we make use of the IIASA Model/scenarios
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.gdp
}
}
\keyword{datasets}
