% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.mort.rates.ctry_ctry}
\alias{raw.mort.rates.ctry_ctry}
\title{raw.mort.rates.ctry_ctry}
\format{
.csv
}
\source{
TM5-FASST, Global Burden of Disease Study 2021 (GBD 2021) Results; Seattle, United States: Institute for Health Metrics and Evaluation (IHME), 2022.; Available from https://vizhub.healthdata.org/gbd-results/.
}
\usage{
raw.mort.rates.ctry_ctry
}
\description{
cause-specific baseline mortalities from stroke, ischemic heart disease (IHD), chronic obstructive pulmonary disease (COPD), acute lower respiratory illness diseases (ALRI), lung cancer (LC), and diabetes mellitus type II (dm) by age and sex and CTRY.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.mort.rates.ctry_ctry
}
}
\keyword{datasets}
