% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.mort.rates}
\alias{raw.mort.rates}
\title{raw.mort.rates}
\format{
.csv
}
\source{
TM5-FASST, v4.7 -- Global Burden of Disease
}
\usage{
raw.mort.rates
}
\description{
cause-specific baseline mortalities from stroke, ischemic heart disease (IHD), chronic obstructive pulmonary disease (COPD), acute lower respiratory illness diseases (ALRI), lung cancer (LC), and diabetes mellitus type II (dm).
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.mort.rates
}
}
\keyword{datasets}
