% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.nuts.pop}
\alias{raw.nuts.pop}
\title{GDP-SSP database}
\format{
.tsv.gz
}
\source{
https://doi.org/10.2908/CENS_21AGR3
}
\usage{
raw.nuts.pop
}
\description{
Population data from Eurostat 2021 by NUTS0, NUTS1, NUTS2, and NUTS3.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.nuts.pop
}
}
\keyword{datasets}
