% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.rr.fusion}
\alias{raw.rr.fusion}
\title{raw.rr.fusion}
\format{
.csv
}
\source{
Lim, S.S., Vos, T., Flaxman, A.D., Danaei, G., Shibuya, K., Adair-Rohani, H., AlMazroa, M.A., Amann, M., Anderson, H.R., Andrews, K.G. and Aryee, M., 2012. A comparative risk assessment of burden of disease and injury attributable to 67 risk factors and risk factor clusters in 21 regions, 1990–2010: a systematic analysis for the Global Burden of Disease Study 2010. The lancet, 380(9859), pp.2224-2260.
}
\usage{
raw.rr.fusion
}
\description{
Relative risk of death attributable to a change in population-weighted mean pollutant concentration using the FUSION model. From Burnett, R.T., Spadaro, J.V., Garcia, G.R. and Pope, C.A., 2022. Designing health impact functions to assess marginal changes in outdoor fine particulate matter. Environmental Research, 204, p.112245.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.rr.fusion
}
}
\keyword{datasets}
