% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.rr.gbd.param}
\alias{raw.rr.gbd.param}
\title{raw.rr.gbd.param}
\format{
.csv
}
\source{
GBD 2018
}
\usage{
raw.rr.gbd.param
}
\description{
Parameters for the estimation of the RR of death attributable to a change in population-weighted mean pollutant concentration, based on the GBD 2018 study.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.rr.gbd.param
}
}
\keyword{datasets}
