% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.rr.param.bur2014}
\alias{raw.rr.param.bur2014}
\title{raw.rr.param.bur2014}
\format{
An object of class \code{data.frame} with 5 rows and 6 columns.
}
\source{
Lim, S.S., Vos, T., Flaxman, A.D., Danaei, G., Shibuya, K., Adair-Rohani, H., AlMazroa, M.A., Amann, M., Anderson, H.R., Andrews, K.G. and Aryee, M., 2012. A comparative risk assessment of burden of disease and injury attributable to 67 risk factors and risk factor clusters in 21 regions, 1990–2010: a systematic analysis for the Global Burden of Disease Study 2010. The lancet, 380(9859), pp.2224-2260.
}
\usage{
raw.rr.param.bur2014
}
\description{
Relative risk of death attributable to a change in population-weighted mean pollutant concentration. From Van Dingenen et al (2018): "RR for PM2:5 exposure is calculated from the integrated exposure-response (IER) functions developed by Burnett et al. (2014) and first applied in the GBD study (Lim et al., 2012)
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.rr.param.bur2014
}
}
\keyword{datasets}
