% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.rr.param.bur2018.with}
\alias{raw.rr.param.bur2018.with}
\title{raw.rr.param.bur2018.with}
\format{
.csv
}
\source{
Burnett, R., Chen, H., Szyszkowicz, M., Fann, N., Hubbell, B., Pope Iii, C.A., Apte, J.S., Brauer, M., Cohen, A., Weichenthal, S. and Coggins, J., 2018. Global estimates of mortality associated with long-term exposure to outdoor fine particulate matter. Proceedings of the National Academy of Sciences, 115(38), pp.9592-9597.
}
\usage{
raw.rr.param.bur2018.with
}
\description{
Parameters for the estimation of the RR of death attributable to a change in population-weighted mean pollutant concentration, with the extreme Chinese cohort, based on the GEMM model from Burnett et al 2018.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.rr.param.bur2018.with
}
}
\keyword{datasets}
