% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.rr.param.gbd2016}
\alias{raw.rr.param.gbd2016}
\title{raw.rr.param.gbd2016}
\format{
.csv
}
\source{
Forouzanfar, M.H., Afshin, A., Alexander, L.T., Anderson, H.R., Bhutta, Z.A., Biryukov, S., Brauer, M., Burnett, R., Cercy, K., Charlson, F.J. and Cohen, A.J., 2016. Global, regional, and national comparative risk assessment of 79 behavioural, environmental and occupational, and metabolic risks or clusters of risks, 1990–2015: a systematic analysis for the Global Burden of Disease Study 2015. The lancet, 388(10053), pp.1659-1724.
}
\usage{
raw.rr.param.gbd2016
}
\description{
Parameters for the estimation of the RR of death attributable to a change in population-weighted mean pollutant concentration, based on the GBD 2016 study.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.rr.param.gbd2016
}
}
\keyword{datasets}
