% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.ssp.data}
\alias{raw.ssp.data}
\title{SSP data}
\format{
.csv
}
\source{
https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome
}
\usage{
raw.ssp.data
}
\description{
Country level population and GDP data per SSP (SSP_database_v9.csv). To be consistent we make use of the IIASA-WIC Model/scenarios
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.ssp.data
}
}
\keyword{datasets}
