% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.twn.pop}
\alias{raw.twn.pop}
\title{Socioeconomics Taiwan}
\format{
.csv
}
\source{
https://tntcat.iiasa.ac.at/SspDb/dsd?Action=htmlpage&page=welcome
}
\usage{
raw.twn.pop
}
\description{
Population and GDP data per SSP (SSP_database_v9.csv). Given that IIASA-WIC does not report data for Taiwan, we use data from "OECD_Env-Growth"
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.twn.pop
}
}
\keyword{datasets}
