% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.urb_incr}
\alias{raw.urb_incr}
\title{Urban Increment}
\format{
.csv
}
\source{
Van Dingenen, R., Dentener, F., Crippa, M., Leitao, J., Marmer, E., Rao, S., Solazzo, E. and Valentini, L., 2018. TM5-FASST: a global atmospheric source-receptor model for rapid impact analysis of emission changes on air quality and short-lived climate pollutants. Atmospheric Chemistry and Physics, 18(21), pp.16173-16211.
}
\usage{
raw.urb_incr
}
\description{
Adjustment factor for primary PM2.5 concentrations (BC and POM), which are assumed to be concentrated in urban areas.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.urb_incr
}
}
\keyword{datasets}
