% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.yll.pm25}
\alias{raw.yll.pm25}
\title{raw.yll.pm25}
\format{
An object of class \code{data.frame} with 56 rows and 8 columns.
}
\source{
TM5-FASST
}
\usage{
raw.yll.pm25
}
\description{
Years of Life Lost (YLLs) to Mortality ratios attributable to PM2.5 exposure
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.yll.pm25
}
}
\keyword{datasets}
