% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Regions}
\alias{Regions}
\title{Regions}
\format{
An object of class \code{data.frame} with 249 rows and 4 columns.
}
\usage{
Regions
}
\description{
Combined regions
\dontrun{
 library(rfasst);
 rfasst::Regions
}
}
\keyword{datasets}
