% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Regions_EUR}
\alias{Regions_EUR}
\title{Regions_EUR}
\format{
An object of class \code{data.frame} with 258 rows and 4 columns.
}
\usage{
Regions_EUR
}
\description{
Combined GCAM-Europe regions
\dontrun{
 library(rfasst);
 rfasst::Regions_EUR
}
}
\keyword{datasets}
