% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Regions_vlt8.2}
\alias{Regions_vlt8.2}
\title{Regions_vlt8.2}
\format{
An object of class \code{data.frame} with 258 rows and 4 columns.
}
\usage{
Regions_vlt8.2
}
\description{
Combined GCAM (version < 8.2)regions
\dontrun{
 library(rfasst);
 rfasst::Regions_vlt8.2
}
}
\keyword{datasets}
