% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rr_resp_o3_GBD2016_high}
\alias{rr_resp_o3_GBD2016_high}
\title{rr_resp_o3_GBD2016_high}
\format{
An object of class \code{numeric} of length 1.
}
\source{
TM5-FASST
\dontrun{
 library(rfasst);
 rfasst::rr_resp_o3_GBD2016_high
}
}
\usage{
rr_resp_o3_GBD2016_high
}
\description{
Relative risk for respiratory disease associated to ozone exposure from GBD2016
}
\keyword{datasets}
