% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selected_pollutants}
\alias{selected_pollutants}
\title{Selected pollutants}
\format{
vector

An object of class \code{character} of length 30.
}
\source{
-
}
\usage{
selected_pollutants

selected_pollutants
}
\description{
Selected pollutants

Pollutants to be considered
\dontrun{
 library(rfasst);
 rfasst::selected_pollutants
}
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::selected_pollutants
}
}
\keyword{datasets}
