% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src.bc}
\alias{src.bc}
\title{SRC BC}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
src.bc
}
\description{
Source-receptor coefficients (SRC) between BC emissions and BC concentration level, pre-computed from a 20% emission reduction of component i  in region xk relative to the base run
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::src.bc
}
}
\keyword{datasets}
