% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src.m6m_nox}
\alias{src.m6m_nox}
\title{src.m6m_nox}
\format{
.csv

.csv
}
\source{
Results from TM5
}
\usage{
src.m6m_nox
}
\description{
Source-receptor coefficients (SRC) between NOx emissions and M6M (O3) concentration level, pre-computed from a 20% emission reduction of component i  in region xk relative to the base run
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::src.m6m_nox
}
}
\keyword{datasets}
