% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src.maize_aot40_so2}
\alias{src.maize_aot40_so2}
\title{src.maize_aot40_so2}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
src.maize_aot40_so2
}
\description{
Source-receptor coefficients (SRC) between SO2 emissions and AOT40 (O3) concentration level for maize, pre-computed from a 20% emission reduction of component i  in region xk relative to the base run
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::src.maize_aot40_so2
}
}
\keyword{datasets}
