% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src.nh4_nox}
\alias{src.nh4_nox}
\title{nh4_nox}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
src.nh4_nox
}
\description{
Source-receptor coefficients (SRC) between NOx emissions and NH4 (PM2.5) concentration level, pre-computed from a 20% emission reduction of component i  in region xk relative to the base run
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::src.nh4_nox
}
}
\keyword{datasets}
