% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src.no3_nh3}
\alias{src.no3_nh3}
\title{no3_nh3}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
src.no3_nh3
}
\description{
Source-receptor coefficients (SRC) between NH3 emissions and NO3 (PM2.5) concentration level, pre-computed from a 20% emission reduction of component i  in region xk relative to the base run
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::src.no3_nh3
}
}
\keyword{datasets}
