% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src.o3_ch4}
\alias{src.o3_ch4}
\title{src.o3_ch4}
\format{
.csv
}
\source{
Fiore, A.M., West, J.J., Horowitz, L.W., Naik, V. and Schwarzkopf, M.D., 2008. Characterizing the tropospheric ozone response to methane emission controls and the benefits to climate and air quality. Journal of Geophysical Research: Atmospheres, 113(D8).
}
\usage{
src.o3_ch4
}
\description{
Source-receptor coefficients (SRC) between CH4 emissions and O3 concentration level, normalized from existing literature (not pre-computed with TM5-FASST)
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::src.o3_ch4
}
}
\keyword{datasets}
