% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src.o3_nmvoc}
\alias{src.o3_nmvoc}
\title{src.o3_nmvoc}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
src.o3_nmvoc
}
\description{
Source-receptor coefficients (SRC) between NMVOC emissions and O3 concentration level, pre-computed from a 20% emission reduction of component i  in region xk relative to the base run
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::src.o3_nmvoc
}
}
\keyword{datasets}
