% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src.pom}
\alias{src.pom}
\title{SRC POM}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
src.pom
}
\description{
Source-receptor coefficients (SRC) between POM emissions and POM concentration level, pre-computed from a 20% emission reduction of component i  in region xk relative to the base run
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::src.pom
}
}
\keyword{datasets}
