% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src.rice_aot40_nox}
\alias{src.rice_aot40_nox}
\title{src.rice_aot40_nox}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
src.rice_aot40_nox
}
\description{
Source-receptor coefficients (SRC) between NOx emissions and AOT40 (O3) concentration level for rice, pre-computed from a 20% emission reduction of component i  in region xk relative to the base run
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::src.rice_aot40_nox
}
}
\keyword{datasets}
