% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src.so4_so2}
\alias{src.so4_so2}
\title{so4_so2}
\format{
.csv
}
\source{
Results from TM5
}
\usage{
src.so4_so2
}
\description{
Source-receptor coefficients (SRC) between SO2 emissions and SO4 (PM2.5) concentration level, pre-computed from a 20% emission reduction of component i  in region xk relative to the base run
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::src.so4_so2
}
}
\keyword{datasets}
