% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TG_KG}
\alias{TG_KG}
\title{TG_KG}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TG_KG
}
\description{
Transform Tg to Kg
\dontrun{
 library(rfasst);
 rfasst::TG_KG
}
}
\keyword{datasets}
