% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vsl_eu_2005_lb}
\alias{vsl_eu_2005_lb}
\title{vsl_eu_2005_lb}
\format{
An object of class \code{numeric} of length 1.
}
\source{
OECD
\dontrun{
 library(rfasst);
 rfasst::vsl_eu_2005_lb
}
}
\usage{
vsl_eu_2005_lb
}
\description{
Lower bound for the Value of Statistical Life (VSL)
}
\keyword{datasets}
