% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vsly_eu_2014}
\alias{vsly_eu_2014}
\title{vsly_eu_2014}
\format{
An object of class \code{numeric} of length 1.
}
\source{
Schlander, M., Schaefer, R. and Schwarz, O., 2017. Empirical studies on the economic value of a Statistical Life Year (VSLY) in Europe: what do they tell us?. Value in Health, 20(9), p.A666.
\dontrun{
 library(rfasst);
 rfasst::vsly_eu_2014
}
}
\usage{
vsly_eu_2014
}
\description{
Base Value of Statistical Life Year (VSLY) in $2014
}
\keyword{datasets}
