% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight.nuts.pop.nuts3}
\alias{weight.nuts.pop.nuts3}
\title{GDP-SSP database}
\format{
.tsv.gz
}
\source{
https://doi.org/10.2908/CENS_21AGR3
}
\usage{
weight.nuts.pop.nuts3
}
\description{
NUTS3 population weight mean by age among all European countries based on Eurostat data 2021 over F, M and T sex.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::weight.nuts.pop.nuts3
}
}
\keyword{datasets}
