% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight.nuts.pop}
\alias{weight.nuts.pop}
\title{GDP-SSP database}
\format{
.tsv.gz
}
\source{
https://doi.org/10.2908/CENS_21AGR3
}
\usage{
weight.nuts.pop
}
\description{
NUTS3 population weights by each European country based on Eurostat data 2021.
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::weight.nuts.pop
}
}
\keyword{datasets}
