% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_health.R
\name{calc_mort_rates}
\alias{calc_mort_rates}
\title{calc_mort_rates}
\source{
https://www.who.int/healthinfo/global_burden_disease/cod_2008_sources_methods.pdf
}
\usage{
calc_mort_rates()
}
\value{
Baseline mortality rates for TM5-FASST regions for all years and causes (ALRI, COPD, LC, IHD, STROKE). The list of countries that form each region and the full name of the region can be found in Table S2.2 in the TM5-FASST documentation paper: Van Dingenen, R., Dentener, F., Crippa, M., Leitao, J., Marmer, E., Rao, S., Solazzo, E. and Valentini, L., 2018. TM5-FASST: a global atmospheric source-receptor model for rapid impact analysis of emission changes on air quality and short-lived climate pollutants. Atmospheric Chemistry and Physics, 18(21), pp.16173-16211.
}
\description{
Get cause-specific baseline mortalities from stroke, ischemic heart disease (IHD), chronic obstructive pulmonary disease (COPD), acute lower respiratory illness diseases (ALRI) and lung cancer (LC).
}
\keyword{Baseline}
\keyword{mortality}
\keyword{rates}
