% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_functions.R
\name{fill_queries}
\alias{fill_queries}
\title{fill_queries}
\usage{
fill_queries(
  prj,
  db_path,
  db_name,
  prj_name,
  scenarios,
  query_path = "inst/extdata",
  queries = "queries_rfasst_nonCO2.xml"
)
}
\arguments{
\item{query_path}{Path to the query file}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{prj:}{current project}

\item{db_path:}{path of the database}

\item{db_name:}{name of the database}

\item{prj_name:}{name of the project}

\item{scenarios:}{name of the scenarios to be considered}
}
\value{
prj containing the nonCO2 emissions by sector
}
\description{
Create a folder to save the datasets and file, in case it does not exist
}
