% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1_emissions_rescale.R
\name{m1_emissions_rescale}
\alias{m1_emissions_rescale}
\title{m1_emissions_rescale}
\usage{
m1_emissions_rescale(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name = NULL,
  prj = NULL,
  rdata_name = NULL,
  scen_name,
  queries = "queries_rfasst.xml",
  final_db_year = 2100,
  saveOutput = T,
  map = F,
  mapIndivPol = F,
  anim = T,
  recompute = F
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{prj}{rgcam loaded project}

\item{rdata_name}{Name of the RData file. It must contain the queries in a list}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{final_db_year}{Final year in the GCAM database (this allows to process databases with user-defined "stop periods")}

\item{saveOutput}{Writes the emission files. By default=T}

\item{map}{Produce the maps. By default=F}

\item{mapIndivPol}{If set to T, it produces the maps for individual pollutants. By default=F}

\item{anim}{If set to T, produces multi-year animations. By default=T}

\item{recompute}{If set to T, recomputes the function output. Otherwise, if the output was already computed once, it uses that value and avoids repeating computations. By default=F}
}
\value{
Emissions per TM5-FASST region (Kg) for all the selected years (all_years). The list of countries that form each region and the full name of the region can be found in Table S2.2 in the TM5-FASST documentation paper: Van Dingenen, R., Dentener, F., Crippa, M., Leitao, J., Marmer, E., Rao, S., Solazzo, E. and Valentini, L., 2018. TM5-FASST: a global atmospheric source-receptor model for rapid impact analysis of emission changes on air quality and short-lived climate pollutants. Atmospheric Chemistry and Physics, 18(21), pp.16173-16211.
}
\description{
Re-scale emissions from GCAM regions to TM5-FASST regions. The function also completes some additional transformation required such as pollutant transformation (e.g. OC to POM) or unit changes (Tg to Kg).
}
\keyword{emissions}
\keyword{module_1,}
\keyword{re-scale}
