% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1_emissions_rescale.R
\name{m1_emissions_rescale}
\alias{m1_emissions_rescale}
\title{m1_emissions_rescale}
\usage{
m1_emissions_rescale(
  db_path,
  query_path,
  db_name,
  prj_name,
  scen_name,
  queries,
  saveOutput = T
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{saveOutput}{Writes the emission files. By default=T}
}
\value{
Emissions per TM5-FASST region for all the selected years (all_years)
}
\description{
Re-scales emissions from GCAM regions to TM5-FASST regions. The function also completes some additional transformation required such as pollutant transformation (e.g. OC to POM) or unit changes (Tg to Kg).
}
\keyword{emissions}
\keyword{module_1,}
\keyword{re-scale}
