% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_conc.R
\name{m2_get_conc_o3}
\alias{m2_get_conc_o3}
\title{m2_get_conc_o3}
\usage{
m2_get_conc_o3(
  db_path,
  query_path,
  db_name,
  prj_name,
  scen_name,
  queries,
  saveOutput = T,
  ch4_o3 = T,
  map = F,
  anim = T
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{saveOutput}{Writes the emission files.By default=T}

\item{ch4_o3}{Includes the CH4 effect on O3 based on Fiore et al (2008).By default=T}

\item{map}{Produce the maps. By default=F}

\item{anim}{If set to T, produces multi-year animations. By default=T}
}
\value{
Produce ozone (O3) levels for each TM5-FASST regions for all years (ppb). The list of countries that form each region and the full name of the region can be found in Table S2.2 in the TM5-FASST documentation paper: Van Dingenen, R., Dentener, F., Crippa, M., Leitao, J., Marmer, E., Rao, S., Solazzo, E. and Valentini, L., 2018. TM5-FASST: a global atmospheric source-receptor model for rapid impact analysis of emission changes on air quality and short-lived climate pollutants. Atmospheric Chemistry and Physics, 18(21), pp.16173-16211.
}
\description{
Produce ozone (O3) concentration levels based on re-scaled emission pathways.
}
\keyword{O3}
\keyword{concentration,}
\keyword{module_2,}
