% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_health.R
\name{m3_get_daly_pm25}
\alias{m3_get_daly_pm25}
\title{m3_get_daly_pm25}
\source{
Institute for Health Metrics and Evaluation (http://www.healthdata.org/)
}
\usage{
m3_get_daly_pm25(
  db_path,
  query_path,
  db_name,
  prj_name,
  scen_name,
  queries,
  ssp = "SSP2",
  saveOutput = T
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{scen_name}{Name of the GCAM scenario to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{ssp}{Set the ssp narrative associated to the GCAM scenario. c("SSP1","SSP2","SSP3","SSP4","SSP5"). By default is SSP2}

\item{saveOutput}{Writes the emission files.By default=T}
}
\value{
Economic damages associated with YLLs attributable to PM2.5 exposure for each TM5-FASST regions for all years
}
\description{
Produce Disability Adjusted Life Years (DALYs) attributable to PM2.5 exposure. See calc_daly_pm for detials on DALY-to-Mortality ratios.
}
\keyword{DALY,}
\keyword{PM2.5,}
\keyword{module_3,}
