% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Percen}
\alias{Percen}
\title{#' Percen}
\format{
An object of class \code{data.frame} with 17490 rows and 6 columns.
}
\source{
RCP database: https://tntcat.iiasa.ac.at/RcpDb/dsd?Action=htmlpage&page=welcome
\dontrun{
 library(rfasst);
 rfasst::Percen
}
}
\usage{
Percen
}
\description{
Percentages to downscale GCAM emissions to country-level
}
\keyword{datasets}
