% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.mort.rates}
\alias{raw.mort.rates}
\title{raw.mort.rates}
\format{
.csv
}
\source{
https://www.who.int/healthinfo/global_burden_disease/cod_2008_sources_methods.pdf
}
\usage{
raw.mort.rates
}
\description{
cause-specific baseline mortalities from stroke, ischemic heart disease (IHD), chronic obstructive pulmonary disease (COPD), acute lower respiratory illness diseases (ALRI) and lung cancer (LC).
}
\examples{
\dontrun{
 library(rfasst);
 rfasst::raw.mort.rates
}
}
\keyword{datasets}
